refactor(controller): address Network Restore review
PR #9124 review
